requirements.txt
# Core dependencies
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Deep learning framework
tensorflow>=2.13.0
# OR use PyTorch-based Mask R-CNN:
# torch>=2.0.0
# torchvision>=0.15.0

# Computer vision libraries
numpy>=1.24.0
Pillow>=10.0.0

# Data handling
pyyaml>=6.0
tqdm>=4.66.0  # Progress bars

# Optional but recommended
matplotlib>=3.7.0  # For visualization and debugging
ipython>=8.12.0    # For interactive development
jupyter>=1.0.0     # Optional: for experimentation

